+++ test/project_types/script/layers/infrastructure/script_service_test.rb
@@ -55,11 +55,6 @@ describe Script::Layers::Infrastructure::ScriptService do
     HERE
   end
 
-  before do
-    ::Script::Layers::Infrastructure::ScriptService::MakeRequest.stubs(:bypass_partners_proxy).returns(false)
-    # script_service.stubs(:bypass_partners_proxy).returns(false)
-  end
-
   describe ".push" do
     let(:script_content) { "(module)" }
     let(:api_key) { "fake_key" }
